pyproject.toml
[project]
name = "notion-nlp"
version = "0.1.0"
description = "A Python library for processing Notion documents with NLP capabilities and hierarchical organization"
authors = [
    {name = "Notion NLP Team", email = "info@example.com"},
]
requires-python = ">=3.11"
license = {text = "MIT"}
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "notion-client>=2.3.0",
    "pydantic>=2.10.6",
    "pytest-mock>=3.14.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "spacy>=3.8.4",
    "streamlit>=1.42.0",
    "trafilatura>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
]
ui = [
    "streamlit>=1.42.0",
]
notebook = [
    "jupyter>=1.1.1",
    "notebook>=7.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["."]
